Mejora en el código en el service de game

+++ games/src/main/java/com/playstore/games/application/GameUseCase.java
@@ -3,27 +3,20 @@ package com.playstore.games.application;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
-import org.apache.commons.io.FilenameUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.cloudinary.Cloudinary;
-import com.cloudinary.utils.ObjectUtils;
 import com.playstore.games.application.exception.CategoryNotFoundException;
 import com.playstore.games.application.exception.GameNotFoundException;
-import com.playstore.games.application.utils.CalculateFinalPrice;
-import com.playstore.games.application.utils.CategoryUtils;
 import com.playstore.games.application.utils.GameUtils;
+import com.playstore.games.application.utils.ImageUtils;
 import com.playstore.games.domain.ECategory;
 import com.playstore.games.domain.Game;
 import com.playstore.games.domain.GameImage;
-import com.playstore.games.infrastructure.dto.GameImageDTO;
 import com.playstore.games.infrastructure.dto.GameRequestDTO;
 import com.playstore.games.infrastructure.dto.GameResponseDTO;
 import com.playstore.games.infrastructure.inputPort.IGameInputPort;
@@ -40,36 +33,15 @@ public class GameUseCase implements IGameInputPort {
     private IGameImageMethod gameImageMethod;
 
     @Autowired
-    private Cloudinary cloudinary;
+    private ImageUtils imageUtils;
 
     @Override
     public GameResponseDTO createGame(GameRequestDTO game, MultipartFile file)
             throws CategoryNotFoundException, IOException {
 
-        BigDecimal finalPrice = CalculateFinalPrice.calculateFinalPrice(game.getPrice(), game.getDiscount());
-        ECategory gameCategory = CategoryUtils.setCategory(game.getCategory());
-        GameImage gameImage = null;
-
-        if (file != null && file.isEmpty() != true) {
-
-            if (file.getSize() > 5000000) {
-                throw new IllegalArgumentException("The image exceeds 5 MB");
-            }
-
-            String originalFilename = file.getOriginalFilename();
-            String extension = FilenameUtils.getExtension(originalFilename);
-
-            if (!extension.equals("jpg") && !extension.equals("jpeg") && !extension.equals("png")) {
-                throw new IllegalArgumentException("The allowed extensions are ‘.jpg’, ‘.jpe’, ‘.png’");
-            }
-
-            Map uploadResult = cloudinary.uploader().upload(file.getBytes(), ObjectUtils.emptyMap());
-            String url = uploadResult.get("url").toString();
-
-            gameImage = GameImage.builder()
-                    .image_url(url)
-                    .build();
-        }
+        BigDecimal finalPrice = GameUtils.calculateFinalPrice(game.getPrice(), game.getDiscount());
+        ECategory gameCategory = GameUtils.setCategory(game.getCategory());
+        GameImage gameImage = imageUtils.fileUpload(file);
 
         Game newGame = gameMethod.save(Game.builder()
                 .title(game.getTitle())
@@ -83,55 +55,19 @@ public class GameUseCase implements IGameInputPort {
                 .gameImage(gameImage)
                 .build());
 
-        GameImageDTO gameImageDTO = (gameImage != null) ? GameImageDTO.builder()
-                .id(newGame.getGameImage().getId())
-                .image_url(newGame.getGameImage().getImage_url())
-                .build() : null;
+        return GameUtils.convertToGameResponseDTO(newGame);
 
-        return GameResponseDTO.builder()
-                .id(newGame.getId())
-                .title(newGame.getTitle())
-                .description(newGame.getDescription())
-                .original_price(newGame.getOriginal_price())
-                .final_price(newGame.getFinal_price())
-                .discount(newGame.getDiscount())
-                .release_date(newGame.getRelease_date())
-                .category(newGame.getCategory().toString())
-                .enabled(newGame.isEnabled())
-                .image(gameImageDTO)
-                .build();
     }
 
     @Override
     public GameResponseDTO updateGame(Long id, GameRequestDTO game, MultipartFile file)
             throws GameNotFoundException, CategoryNotFoundException, IOException {
         Game gameDB = gameMethod.findById(id);
-        Long imgId = gameDB.getGameImage().getId();
+        Long imgId = (gameDB.getGameImage() != null) ? gameDB.getGameImage().getId() : null;
 
-        BigDecimal finalPrice = CalculateFinalPrice.calculateFinalPrice(game.getPrice(), game.getDiscount());
-        ECategory eCategory = CategoryUtils.setCategory(game.getCategory());
-        GameImage gameImage = null;
-
-        if (file != null && file.isEmpty() != true) {
-
-            if (file.getSize() > 5000000) {
-                throw new IllegalArgumentException("The image exceeds 5 MB");
-            }
-
-            String originalFilename = file.getOriginalFilename();
-            String extension = FilenameUtils.getExtension(originalFilename);
-
-            if (!extension.equals("jpg") && !extension.equals("jpeg") && !extension.equals("png")) {
-                throw new IllegalArgumentException("The allowed extensions are ‘.jpg’, ‘.jpe’, ‘.png’");
-            }
-
-            Map uploadResult = cloudinary.uploader().upload(file.getBytes(), ObjectUtils.emptyMap());
-            String url = uploadResult.get("url").toString();
-
-            gameImage = GameImage.builder()
-                    .image_url(url)
-                    .build();
-        }
+        BigDecimal finalPrice = GameUtils.calculateFinalPrice(game.getPrice(), game.getDiscount());
+        ECategory eCategory = GameUtils.setCategory(game.getCategory());
+        GameImage gameImage = imageUtils.fileUpload(file);
 
         gameDB.setTitle(game.getTitle());
         gameDB.setDescription(game.getDescription());
@@ -142,28 +78,14 @@ public class GameUseCase implements IGameInputPort {
         gameDB.setCategory(eCategory);
         if (gameImage != null) {
             gameDB.setGameImage(gameImage);
-            gameImageMethod.deleteById(imgId);
+            if (imgId != null) {
+                gameImageMethod.deleteById(imgId);
+            }
         }
 
         Game gameUpdated = gameMethod.save(gameDB);
 
-        GameImageDTO gameImageDTO = (gameImage != null) ? GameImageDTO.builder()
-                .id(gameUpdated.getGameImage().getId())
-                .image_url(gameUpdated.getGameImage().getImage_url())
-                .build() : null;
-
-        return GameResponseDTO.builder()
-                .id(gameUpdated.getId())
-                .title(gameUpdated.getTitle())
-                .description(gameUpdated.getDescription())
-                .original_price(gameUpdated.getOriginal_price())
-                .final_price(gameUpdated.getFinal_price())
-                .discount(gameUpdated.getDiscount())
-                .release_date(gameUpdated.getRelease_date())
-                .category(gameUpdated.getCategory().toString())
-                .enabled(gameUpdated.isEnabled())
-                .image(gameImageDTO)
-                .build();
+        return GameUtils.convertToGameResponseDTO(gameUpdated);
 
     }
